requirements.txt
Django==1.7.3
django-braces==1.4.0
django-cors-headers==1.0.0
django-facebook==6.0.2
django-ipware==0.1.0
django-oauth-toolkit==0.7.2
django-social-auth==0.7.28
facebook-sdk==0.4.0
facepy==0.8.2
httplib2==0.9
ipdb==0.8
ipython==2.4.0
MySQL-python==1.2.5
oauth2==1.5.211
oauthlib==0.7.2
Pillow==2.7.0
PyJWT==0.4.1
python-openid==2.2.5
python-social-auth==0.2.1
pytz==2014.10
requests==2.5.1
requests-oauthlib==0.4.2
six==1.9.0
Unidecode==0.4.17
